+++ src/main/java/com/example/algoproject/github/dto/request/CreateWebhook.java
@@ -1,4 +1,4 @@
-package com.example.algoproject.study.dto.request;
+package com.example.algoproject.github.dto.request;
 
 import lombok.Data;
 
@@ -6,10 +6,6 @@ import javax.validation.constraints.NotNull;
 
 @Data
 public class CreateWebhook {
-
-//    @NotNull
-//    private String[] events = {"member", "repository", "push"};
-
     @NotNull
     private String[] events;
