+++ src/main/java/com/example/algoproject/study/service/StudyService.java
@@ -3,16 +3,15 @@ package com.example.algoproject.study.service;
 import com.example.algoproject.belongsto.service.BelongsToService;
 import com.example.algoproject.belongsto.domain.BelongsTo;
 import com.example.algoproject.errors.exception.badrequest.AlreadyExistMemberException;
-import com.example.algoproject.errors.exception.badrequest.AlreadyExistRepositoryNameException;
 import com.example.algoproject.errors.exception.badrequest.SameNameException;
 import com.example.algoproject.errors.exception.badrequest.SameUserException;
 import com.example.algoproject.errors.exception.forbidden.NotLeaderUserException;
 import com.example.algoproject.errors.exception.forbidden.StudyAuthException;
 import com.example.algoproject.errors.exception.notfound.NotExistMemberException;
-import com.example.algoproject.errors.exception.notfound.NotExistRepositoryException;
 import com.example.algoproject.errors.exception.notfound.NotExistStudyException;
 import com.example.algoproject.errors.response.CommonResponse;
 import com.example.algoproject.errors.response.ResponseService;
+import com.example.algoproject.github.service.GithubService;
 import com.example.algoproject.study.domain.Study;
 import com.example.algoproject.study.dto.request.*;
 import com.example.algoproject.study.dto.response.StudyInfo;
@@ -24,12 +23,8 @@ import com.example.algoproject.user.dto.UserInfo;
 import com.example.algoproject.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -41,16 +36,11 @@ import java.util.Objects;
 @Service
 public class StudyService {
 
-    @Value("${baseUrl}")
-    private String url;
-
-    @Value("${baseUrl2}")
-    private String url2;
-
     private final StudyRepository studyRepository;
     private final UserService userService;
     private final BelongsToService belongsToService;
     private final ResponseService responseService;
+    private final GithubService githubService;
 
     @Transactional
     public CommonResponse create(CustomUserDetailsVO cudVO, CreateStudy request) {
@@ -61,17 +51,13 @@ public class StudyService {
         log.info("leader name: " + leader.getName());
 
         // 팀장의 github 이름으로 repoName 이 이름인 레포지토리 생성
-        Map<String, Object> response = createRepositoryResponse(leader, request.getRepoName());
-
-        String[] events1 = {"member", "repository"};
-        String[] events2 = {"push"};
+        Map<String, Object> response = githubService.createRepository(leader, request.getRepoName());
 
         // 스터디에 해당하는 Repository 의 webhook 생성
-        createWebhookResponse(leader, request.getRepoName(), this.url, events1);
-        createWebhookResponse(leader, request.getRepoName(), this.url2, events2);
+        githubService.createWebhook(leader, request.getRepoName());
 
         Study study = new Study(response.get("id").toString(), request.getStudyName(), cudVO.getUsername(), response.get("name").toString(), response.get("html_url").toString());
-        save(study);
+        studyRepository.save(study);
 
         // 스터디 생성시 팀장을 스터디 멤버에 추가
         belongsToService.save(new BelongsTo(leader, study));
@@ -93,7 +79,7 @@ public class StudyService {
 
         // 스터디의 이름을 변경
         study.setName(request.getName());
-        save(study);
+        studyRepository.save(study);
 
         return responseService.getSingleResponse(new StudyInfo(study, getMemberInfos(getMembers(study))));
     }
@@ -126,7 +112,7 @@ public class StudyService {
         checkLeader(user, study);
 
         // github API 로 저장소 삭제 요청
-        deleteRepositoryResponse(user, study);
+        githubService.deleteRepository(user, study);
 
         // 데이터베이스에서 스터디와 연관된 belongto들을 삭제
         belongsToService.deleteAllByStudy(study);
@@ -159,7 +145,7 @@ public class StudyService {
             throw new AlreadyExistMemberException();
         } catch (NotExistMemberException ex) {
             // leader 가 github 에서 member 에게 study 레포지토리로 contributor 초대를 보냄
-            addContributorResponse(leader, member, study);
+            githubService.addContributor(leader, member, study);
         }
         return responseService.getSuccessResponse();
     }
@@ -185,7 +171,7 @@ public class StudyService {
         checkMember(member, study);
 
         // leader 가 github 에서 member 에게 study 레포지토리로 contributor 초대를 보냄
-        removeContributorResponse(leader, member, study);
+        githubService.removeContributor(leader, member, study);
 
         return responseService.getSuccessResponse();
     }
@@ -205,51 +191,46 @@ public class StudyService {
     }
 
     @Transactional
-    public void webhook(Map<String, Object> response) {
-
+    public void repoWebhook(Map<String, Object> response) {
         // Webhook 이 왔을 때 경우(contributor, repository)를 분리하여 method 실행
         Map<String, Object> repoMap = (Map<String, Object>) response.get("repository");
         Study study = findById(repoMap.get("id").toString());
 
-        // contributor
-        if (response.containsKey("member")) {
-            Map<String, Object> memberMap = (Map<String, Object>) response.get("member");
-            User member = userService.findById(memberMap.get("id").toString());
-
-            // contributor 추가됐을 때에 데이터베이스에 추가
-            if (response.get("action").equals("added")) {
-                log.info("member: {} added to study: {}", member.getName(), study.getName());
-                belongsToService.save(new BelongsTo(member, study));
-            }
-
-            // contributor 삭제됐을 때에 데이터베이스에서 삭제
-            else {
-                log.info("member: {} removed to study: {}", member.getName(), study.getName());
-                belongsToService.deleteByStudyAndMember(study, member);
-            }
+        // 스터디 레포지토리의 이름 변경 됐을 때에 데이터베이스의 레포이름 변경
+        if (response.get("action").equals("renamed")) {
+            log.info("study: {}'s repository renamed to {}", study.getName(), repoMap.get("name").toString());
+            study.setRepositoryName(repoMap.get("name").toString());
+            studyRepository.save(study);
         }
 
-        // repository
+        // 스터디 레포지토리의 삭제됐을 때에 데이터베이스의 스터디 삭제
         else {
-            // 스터디 레포지토리의 이름 변경 됐을 때에 데이터베이스의 레포이름 변경
-            if (response.get("action").equals("renamed")) {
-                log.info("study: {}'s repository renamed to {}", study.getName(), repoMap.get("name").toString());
-                study.setRepositoryName(repoMap.get("name").toString());
-                save(study);
-            }
-
-            // 스터디 레포지토리의 삭제됐을 때에 데이터베이스의 스터디 삭제
-            else {
-                log.info("study: {}'s repository is deleted...Delete study at database, too", study.getName());
-                belongsToService.deleteAllByStudy(study);
-                studyRepository.delete(study);
-            }
+            log.info("study: {}'s repository is deleted...Delete study at database, too", study.getName());
+            belongsToService.deleteAllByStudy(study);
+            studyRepository.delete(study);
         }
     }
 
     @Transactional
-    public void save(Study study) {
-        studyRepository.save(study);
+    public void memberWebhook(Map<String, Object> response) {
+        // Webhook 이 왔을 때 경우(contributor, repository)를 분리하여 method 실행
+        Map<String, Object> repoMap = (Map<String, Object>) response.get("repository");
+        Map<String, Object> memberMap = (Map<String, Object>) response.get("member");
+
+        Study study = findById(repoMap.get("id").toString());
+        User member = userService.findById(memberMap.get("id").toString());
+
+        // contributor 추가됐을 때에 데이터베이스에 추가
+        if (response.get("action").equals("added")) {
+            log.info("member: {} added to study: {}", member.getName(), study.getName());
+            belongsToService.save(new BelongsTo(member, study));
+        }
+
+        // contributor 삭제됐을 때에 데이터베이스에서 삭제
+        else {
+            log.info("member: {} removed to study: {}", member.getName(), study.getName());
+            belongsToService.deleteByStudyAndMember(study, member);
+        }
     }
 
     @Transactional(readOnly = true)
@@ -280,103 +261,6 @@ public class StudyService {
     //
     // private methods
     //
-
-    private Map<String, Object> createRepositoryResponse(User leader, String repoName) {
-
-        HttpHeaders headers = makeHeader(leader);
-        CreateRepository request = new CreateRepository();
-        request.setName(repoName);
-        request.setAuto_init(true);
-        HttpEntity<CreateRepository> entity = new HttpEntity<>(request, headers);
-        RestTemplate restTemplate = new RestTemplate();
-
-        try {
-            ResponseEntity<Map<String, Object>> response = restTemplate.exchange(
-                    "https://api.github.com/user/repos",
-                    HttpMethod.POST,
-                    entity,
-                    new ParameterizedTypeReference<>() {
-                    });
-            return response.getBody();
-        } catch (RuntimeException ex) {
-            throw new AlreadyExistRepositoryNameException();
-        }
-    }
-
-    private void deleteRepositoryResponse(User leader, Study study) {
-
-        HttpHeaders headers = makeHeader(leader);
-        HttpEntity<String> entity = new HttpEntity<>(headers);
-        RestTemplate restTemplate = new RestTemplate();
-
-        try {
-            restTemplate.exchange(
-                    "https://api.github.com/repos/" + leader.getName() + "/" + study.getRepositoryName(),
-                    HttpMethod.DELETE,
-                    entity,
-                    new ParameterizedTypeReference<>() {
-                    });
-        } catch (RuntimeException ex) {
-            throw new NotExistRepositoryException();
-        }
-    }
-
-    private void createWebhookResponse(User leader, String repoName, String webhookUrl, String[] events) {
-        HttpHeaders headers = makeHeader(leader);
-        CreateWebhook request = new CreateWebhook(events);
-        request.getConfig().setUrl(webhookUrl);
-        HttpEntity<CreateWebhook> entity = new HttpEntity<>(request, headers);
-        RestTemplate restTemplate = new RestTemplate();
-        restTemplate.exchange(
-                "https://api.github.com/repos/" + leader.getName() + "/" + repoName + "/hooks",
-                HttpMethod.POST,
-                entity,
-                new ParameterizedTypeReference<>() {
-                });
-    }
-
-    private void addContributorResponse(User leader, User member, Study study) {
-
-        HttpHeaders headers = makeHeader(leader);
-
-        AddContributor request = new AddContributor();
-        request.setPermission("admin");
-
-        HttpEntity<AddContributor> entity = new HttpEntity<>(request, headers);
-
-        RestTemplate restTemplate = new RestTemplate();
-
-        ResponseEntity<Map<String, Object>> response = restTemplate.exchange(
-                "https://api.github.com/repos/" + leader.getName() + "/" + study.getRepositoryName() + "/collaborators/" + member.getName(),
-                HttpMethod.PUT,
-                entity,
-                new ParameterizedTypeReference<>() {
-                });
-        if(response.getStatusCodeValue() == 204)
-            throw new AlreadyExistMemberException();
-    }
-
-    private void removeContributorResponse(User leader, User member, Study study) {
-
-        HttpHeaders headers = makeHeader(leader);
-        HttpEntity<Object> entity = new HttpEntity<>(headers);
-        RestTemplate restTemplate = new RestTemplate();
-
-        restTemplate.exchange(
-                "https://api.github.com/repos/" + leader.getName() + "/" + study.getRepositoryName() + "/collaborators/" + member.getName(),
-                HttpMethod.DELETE,
-                entity,
-                new ParameterizedTypeReference<>() {
-                });
-    }
-
-    private HttpHeaders makeHeader(User owner) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("User-Agent", "api-test");
-        headers.add("Authorization", "token " + owner.getAccessToken());
-        headers.add("Accept", "application/vnd.github.v3+json");
-        return headers;
-    }
 
     private List<User> getMembers(Study study) {
         List<User> members = new ArrayList<>();
